requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
pytz==2022.2
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
